pyproject.toml
[project]
name = "bale-bot"
version = "0.1.0"
description = "A bot that looks for specific Telegram channel messaegs and send them to the specified Bale channel"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp==3.9.1",
    "aiosignal==1.3.2",
    "asttokens==3.0.0",
    "asyncio==3.4.3",
    "attrs==25.3.0",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "cryptography==45.0.4",
    "decorator==5.2.1",
    "executing==2.2.0",
    "frozenlist==1.7.0",
    "idna==3.10",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "matplotlib-inline==0.1.7",
    "multidict==6.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyaes==1.6.1",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygments==2.19.2",
    "python-bale-bot==2.5.0",
    "python-dotenv>=1.2.1",
    "requests==2.32.4",
    "rsa==4.9.1",
    "setuptools==69.0.3",
    "stack-data==0.6.3",
    "telethon==1.40.0",
    "traitlets==5.14.3",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "yarl==1.20.1",
]
